rename and move ultrafeed spotlight item
